% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_renamer.R
\name{col_fixer}
\alias{col_fixer}
\title{Name standardiser}
\usage{
col_fixer(dat)
}
\arguments{
\item{dat}{the \code{data.frame} for which you want to pretty up the column names.}
}
\value{
Returns a \code{data.frame} with the nice column names.
}
\description{
All this does is makes the names the same, no matter if they have weird spaces, characters, or brackets. Note that it does not account for differences in spaces or in capitalisation.
}
