% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/does_station_have_hourly_data.R
\name{does_have_hourly_data}
\alias{does_have_hourly_data}
\title{Does the station have hourly data}
\usage{
does_have_hourly_data(station_ID)
}
\arguments{
\item{station_ID}{the station ID}
}
\value{
\code{no_hourly_data} gives NULL when nothing is wrong, and an "error code" (read a string that will then be used as an error code in another function) when something is wrong.
}
\description{
This function checks if the imputed station ID has any hourly data. This function assumes that the station ID does indeed exist, as it is for use of me and only me, I did not idiot proof it, although maybe I should given that I am an idiot. So it turns out that I AM an idiot, so I added the check.
}
