% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_all_one_station.R
\name{download_all_one_station}
\alias{download_all_one_station}
\title{Full Station Data Downloader}
\usage{
download_all_one_station(
  station_name = NULL,
  station_ID = NULL,
  testing = FALSE,
  temporary = FALSE
)
}
\arguments{
\item{station_name}{the station whose data you want downloaded. This is not necessary to have, but it is nice.}

\item{station_ID}{the station ID of the station whose data you want downloaded. If left empty, will be found using the station name.}

\item{testing}{this determines if the url will actually be downloaded, or if we are just pretending for the sake of both my sanity and my computer. Defaults to FALSE.}

\item{temporary}{this determines if the files are temporary or if they are saved to the computer. Defaults to FALSE.}
}
\value{
Creates a data frame of all of a single stations hourly data across all of the available years if \code{temporary} is set to TRUE, and saves one csv file per month for each available month to your computer if \code{temporary} is set to FALSE.
}
\description{
This function downloads all of the hourly data from a single station.
}
\examples{
kenora_data <- download_all_one_station(station_ID = 3959, temporary = TRUE)
str(kenora_data)

}
