% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_all_of_the_files.R
\name{download_all_station_data}
\alias{download_all_station_data}
\title{Download All Station Data}
\usage{
download_all_station_data(
  testing = FALSE,
  limit_for_testing_purposes,
  temporary = FALSE
)
}
\arguments{
\item{testing}{this determines if the url will actually be downloaded, or if we are just pretending for the sake of both my sanity and my computer. Defults to FALSE.}

\item{limit_for_testing_purposes}{because I want to show that it actually works, you can limit the number of stations that you want, and it will pick that many random stations to download in their entirety.}

\item{temporary}{this determines if the files are temporary or if they are saved to the computer. Defaults to FALSE.}
}
\value{
if \code{temporary} is set to TRUE, this (theoretically) returns a giant data frame of all of the hourly station data from all available years and all available stations. Or apparently as a list of \code{data.frame}s, which I'm not sure why it's doing that and am too tired to figure out. If \code{temporary} is set to FALSE, those files will be downloaded in subdirectories named after their associated station.
}
\description{
This function downloads all of the hourly station data across all avalible stations for every year that data is avalible.
}
\examples{
full_station_data <- download_all_station_data(limit_for_testing_purposes = 2, temporary = TRUE)
str(full_station_data)
}
