% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_the_station_ids_and_names_match.R
\name{matching_station_names_and_ids}
\alias{matching_station_names_and_ids}
\title{Station Name and ID checker}
\usage{
matching_station_names_and_ids(station_ID, station_name)
}
\arguments{
\item{station_ID}{the station ID}

\item{station_name}{the station name}
}
\value{
NULL if nothing is wrong and an error code when something has gone wrong.
}
\description{
Function that makes sure that, when both the station name and ID are provided, the station and ID match.
}
