% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_missing_check.R
\name{miss_id}
\alias{miss_id}
\title{Missing station ID checker}
\usage{
miss_id(station_ID = NULL)
}
\arguments{
\item{station_ID}{the station ID}
}
\value{
missing_id
}
\description{
This is a function that checks to make sure the user inserted a station ID because I really don't want to write this code a hundred times.
}
