% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app_2_barcode_looking_one.R
\name{missing_data_visualiser_barcodes}
\alias{missing_data_visualiser_barcodes}
\title{Interactive graphs for visulising missing data}
\usage{
missing_data_visualiser_barcodes(dat)
}
\arguments{
\item{dat}{a \code{data.frame} of Environment Canada data}
}
\value{
a shiny app that you can interact with to help compare missingness of data across stations.
}
\description{
Makes an interactive barcoode looking graph that shows where and which data point are missing.Just like the other shiny app,  I for some reason could not get my checks to work, so just don't be an idiot and use the data from the database.
}
