% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app_1.R
\name{missing_data_visualiser_heatmap}
\alias{missing_data_visualiser_heatmap}
\title{Interactive heatmap for visulising missing data}
\usage{
missing_data_visualiser_heatmap(dat)
}
\arguments{
\item{dat}{a \code{data.frame} of Environment Canada data}
}
\value{
a shiny app that you can interact with to help compare missingness of data across stations.
}
\description{
Makes an interactive heatmap that shows where and which data point are missing. I for some reason could not get my checks to work, so just don't be an idiot and use the data from the database.
}
