% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_reader_and_col_name_fixer.R
\name{reader_and_name_fixer}
\alias{reader_and_name_fixer}
\title{csv reader and column name fixer}
\usage{
reader_and_name_fixer(csv)
}
\arguments{
\item{csv}{the csv that you want read in with nice column names}
}
\value{
your read in csv file with fixed names
}
\description{
All this does is read in your csv file and gets rid of special characters so that it's easier to work with and/or make an SQLite database out of. I use this a lot, so maybe it'll be helpful to you, user, because the way Environment Canada does thing make me frustrated.
}
