% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_helper.R
\name{station_helper}
\alias{station_helper}
\title{Station Helper}
\usage{
station_helper(province, available_year, starting_letter)
}
\arguments{
\item{province}{the province of interest}

\item{available_year}{the year of interest}

\item{starting_letter}{the starting letter that you are looking for}
}
\value{
A data frame that contains all of the stations that fit the inputted criteria. This \code{data.frame} contains the station name, station ID, province, and the first and last year of hourly data collection.
}
\description{
This function helps you sort through the metadata to find the stations that fit a certain criteria. This function can help you find stations based on their Province, the years for which they have hourly data, and the starting letter(s) of the station name.
}
\examples{
station_helper(province = "Ontario", available_year = 1999,starting_letter = "A")

}
