% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_id_finder.R
\name{station_id_finder}
\alias{station_id_finder}
\title{Station ID finder}
\usage{
station_id_finder(station_name)
}
\arguments{
\item{station_name}{}
}
\value{
station_ID
}
\description{
This function will find your station ID based on the station name and will tell you if there are multiple stations that have the same name.
}
