% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_station_one_year.R
\name{download_all_one_station_one_year}
\alias{download_all_one_station_one_year}
\title{csv downloader, one year, one station}
\usage{
download_all_one_station_one_year(
  station_name = NULL,
  station_ID = NULL,
  year,
  temporary = FALSE,
  testing = FALSE
)
}
\arguments{
\item{station_name}{\code{station_name}; the station name}

\item{station_ID}{\code{station_ID}; the station ID}

\item{year}{\code{year}; the year}
}
\value{
dat \code{dat}; a list containing all the downloaded csv files
}
\description{
Function that downloads hourly data of one year from one station.
}
