% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_one_csv_checks.R
\name{download_just_one_csv_with_checks}
\alias{download_just_one_csv_with_checks}
\title{csv downloader}
\usage{
download_just_one_csv_with_checks(
  station_name = NULL,
  station_ID = NULL,
  year,
  month,
  temporary = FALSE,
  testing = FALSE
)
}
\arguments{
\item{station_name}{\code{station_name}; the station name}

\item{station_ID}{\code{station_ID}; the station ID}

\item{year}{\code{year}; the year}

\item{month}{\code{month}; the month}
}
\value{
a \code{data.frame} that has the hourly data from one month from one year from one station.
}
\description{
Function that downloads one file from one year from one station
}
