% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_SQLite_database.R
\name{make_database}
\alias{make_database}
\title{SQLite Database maker}
\usage{
make_database(file_path_to_csv_files, database_name)
}
\arguments{
\item{file_path_to_csv_files}{the directory in which your csv files are stored}

\item{database_name}{the name you want to give your database. This should end in "sqlite".}
}
\value{
Makes an SQLite database on your computer in your working directory. In this database is the station metadata as well as all of the observations from the csv files provided.
}
\description{
SQLite Database maker
}
