% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_name_finder.R
\name{station_name_finder}
\alias{station_name_finder}
\title{Station Name Finder}
\usage{
station_name_finder(station_ID)
}
\arguments{
\item{station_ID}{}
}
\value{
Give you the station name associated with the entered station ID.
}
\description{
That's it. That's all it does. It finds your station name based on the station ID.
}
